refactor: reorganize project file structure & Extract microphone & connection logic into custom hook
