+++ src/components/Hero.jsx
@@ -6,7 +6,7 @@ import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
-import { HERO_TEXT } from '../constants';
+import { HERO_TEXT } from '../constant/hero';
