+++ src/service/realtimeAPI/stopConnection.js
