+++ src/utils/helper_func.js
